Tidying up some code
